# chore: Optimized DB queries