Still use the old (very fast) query to get most of the entries. For the entries that are missing, look for them in the archived tables (this query is much slower)

I used SQL `EXPLAIN ANALYSE` to compare the queries and:
- the query that uses the `valid_to.eq(MAX_TS)` takes:
````
Planning Time: 0.142 ms     
Execution Time: 0.027 ms    
````
- the one that uses the `distinct` takes:
````
Planning Time: 10.915 ms  
Execution Time: 0.738 ms  
````